Removed placeholder config and moved help module to skybeard
