--- config.py
@@ -1,1 +1,0 @@
-#placeholder file for auto-doc generation
